requirements.txt
akshare>=1.16.0
jsonschema>=4.22.0
markdown>=3.6
openai>=1.55.0
pandas>=2.2.0
playwright>=1.48.0
python-dotenv>=1.0.1
streamlit>=1.40.0
xtquant>=250516.1.1
langgraph>=1.0.0
langchain-core>=0.3.0
pytest>=8.3.0

# FastAPI backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
celery>=5.4.0
redis>=5.2.0
aiosqlite>=0.20.0
croniter>=5.0.0
requests>=2.32.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
python-multipart>=0.0.12
slowapi>=0.1.9

# Optional PDF backend used by services/report/pdf_builder.py.
# The default pipeline uses Playwright/Chromium for PDF export.
weasyprint>=63.0
